Draft: Don't always pass `-undefined dynamic_lookup` when linking Mach-Os
As the now-removed comment said, this is only justified for the cyclic
dependencies of the RTS. For everything else we should just pass the
dependencies to the linker so it can audit that after dynamic loading
all symbols should be defined.

But now we specify `-undefined dynamic_lookup` in the RTS's linker
flags, so that justification no longer applies. Let's remove this and
see what happens.