Add a new I/O manager based on poll()
This is a proof of concept I/O manager, to show how to add new ones
neatly, using the ClosureTable and TimeoutQueue infrastructure.

It uses the old unix poll() API, so it is of course limited in
performance by that, but it should have the benefit of wide
compatibility. Also we neatly avoid a name clash with the existing
select() I/O manager.

Compared to the select() I/O manager:
1. beause it uses poll() it is not limited to 1024 file descriptors
   (but it's still O(n) so don't expect great performance);
2. it should have much faster threadDelay (when using it in lots of
   threads at once) because it's based on the new TimeoutQueue which is
   O(log n) rather than O(n).

Some of the code related to timers/timouts is put into a shared module
rts/posix/Timeout.{h,c} since it is intended to be shared with other
similar I/O managers.